chore: Do not require dexdump to install package
